refactor: move project_user mutations
