+++ docs/sdk/project_user.md
@@ -2,3 +2,6 @@
 
 ## Queries
 ::: kili.queries.project_user.__init__.QueriesProjectUser
+
+## Mutations
+::: kili.mutations.project_user.__init__.MutationsProjectUser
